refactor: optimize test file organization
- Remove unused starter.test.ts file
- Merge starter-basic.test.ts and starter-env.test.ts into basic.test.ts since they share the same fixture root (examples/starter)
- Remove 'starter-' prefix from all test files for cleaner naming:
  - starter-no-ssr.test.ts → no-ssr.test.ts
  - starter-react-compiler.test.ts → react-compiler.test.ts
  - starter-base.test.ts → base.test.ts
  - starter-module-runner.test.ts → module-runner.test.ts
  - starter-render-built-url.test.ts → render-built-url.test.ts

Now only truly independent test files (different fixture roots) are split, allowing for optimal parallelization while maintaining logical grouping.
