+++ packages/plugin-rsc/e2e/react-compiler.test.ts
