--- packages/plugin-rsc/e2e/starter.test.ts
@@ -1,12 +1,0 @@
-// This file is kept for backward compatibility
-// All tests have been split into separate files for better parallelization:
-// - starter-basic.test.ts
-// - starter-no-ssr.test.ts
-// - starter-env.test.ts
-// - starter-react-compiler.test.ts
-// - starter-base.test.ts
-// - starter-module-runner.test.ts
-// - starter-render-built-url.test.ts
-
-// To run all starter tests, use:
-// pnpm test-e2e starter-basic starter-no-ssr starter-env starter-react-compiler starter-base starter-module-runner starter-render-built-url
